Refactor Storage to global
