+++ controller/scripts.go
@@ -20,11 +20,11 @@ package controller
 
 import (
 	"github.com/dapperlabs/flow-playground-api/blockchain"
+	"github.com/dapperlabs/flow-playground-api/server/storage"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 
 	"github.com/dapperlabs/flow-playground-api/model"
-	"github.com/dapperlabs/flow-playground-api/storage"
 )
 
 type Scripts struct {
