+++ controller/transactions.go
@@ -21,7 +21,7 @@ package controller
 import (
 	"github.com/dapperlabs/flow-playground-api/blockchain"
 	"github.com/dapperlabs/flow-playground-api/model"
-	"github.com/dapperlabs/flow-playground-api/storage"
+	"github.com/dapperlabs/flow-playground-api/server/storage"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 )
